Bug Fix: moving record manip buttons to E or W didn't lay out right
In the table widget with record manipulation buttons, by default, the
record manipulation buttons appear on the SOUTH side of the table
widget. The developer has the option to move those buttons to any of the
four sides of the table, and the intention was that the six buttons
would be laid out vertically if they are on the EAST or WEST side of the
table, whereas they would be laid out horizontally on the SOUTH or NORTH
side of the table. However, the buttons were laid out horizontally on
all sides. 

The class has been modified to provide the intended behavior.