Fix spacer between table action buttons
Previously, when oriented vertically, the space that appears on
horizontal orientation of the buttons between Delete and MoveToTop did
not appear. Further, if the table layout height was not controlled by
the minimum height of the buttons, the Add and Delete buttons would
appear at the top of the table and the remaining buttons would appear at
the bottom of the table.

This issue has been fixed so that the buttons lay out as intended.